Q and Qu problem fixed.
Change the way showing cells: Instead of creating new cells and replace
old ones, only 16 cells are produced this time. What we need to do is
just updating the letter information on it.

We can refresh the board, which might be useful for left, right, up,
down movement for multigame.